requirements-tests.txt
coverage>=6.0
pytest>=7.0
pytest-cov>=3.0
pytest-flask>=1.2.0
mock>=4.0
flake8>=4.0
autopep8>=1.6
blinker>=1.5